Rewrite sound engine using 16-bit sound

+++ src/main/java/pro/esteps/jsynth/contract/SoundProducer.java
@@ -1,5 +1,5 @@
 package pro.esteps.jsynth.contract;
 
 public interface SoundProducer {
-    public byte[] getSoundChunk();
+    public short[] getSoundChunk();
 }
